feat EmotionCaptureActivity : create main activity
