+++ app/src/main/res/layout/emotioncapture_nav_header.xml
