+++ app/src/main/res/values/strings.xml
@@ -63,4 +63,5 @@
 
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="emotion_capture">Emotie Captatie</string>
 </resources>
